+++ Project_TruyenVN/TruyenVNAPI/DTO/ChapterDTO.cs
@@ -6,7 +6,7 @@ namespace TruyenVNAPI.DTO
 {
     public class ChapterDTO
     {
-        public double chapter_number { get; set; }
+        public int chapter_number { get; set; }
         public int story_id { get; set; }
         public string title { get; set; }
         public string content { get; set; }
